usecases/mwaa-snowflake-integration/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.6.3/constraints-3.10.txt"
snowflake-connector-python
apache-airflow-providers-snowflake
snowflake-sqlalchemy
